Wave 2D

Auxiliary

Introduction to MaxwellWave

Maxwell Wave

Derivation Partial Differentaial Equation

Boundary Condition

Results

Conclusion and Possible Extensions

Running the Project Examples

Wave 2D

Wave 3D

Nonlinear Chi3 Wave 3D

MaxwellWave.Wave2D.update_u! — Function

@parallel_indices (ix,iy) function update_u(u,v,dt)

Computes the update rule for the scalar field through the equation:

\[u^{j+1} = u^{j} + dt*v^{j+\frac{1}{2}}\]

Important to call the kernel with size(u)-2

MaxwellWave.Wave2D.update_v_nabla2! — Function

@parallel_indices (ix,iy) function update_v!(u,v,dt,_dx,_dy,c2)

Computes the update rule for the velocity of the scalar field through the equation:

\[v^{j+\frac{1}{2}} = v^{j-\frac{1}{2}} + dt*c^2 \nabla^2 u^{j}\]

MaxwellWave.Auxiliary.load_array

MaxwellWave.Auxiliary.save_array

MaxwellWave.Wave2D.update_u!

MaxwellWave.Wave2D.update_v_nabla2!

MaxwellWave.Wave3D.compute_vecv_abs_xy_left!

MaxwellWave.Wave3D.compute_vecv_abs_xy_right!

MaxwellWave.Wave3D.compute_vecv_abs_xz_left!

MaxwellWave.Wave3D.compute_vecv_abs_xz_right!

MaxwellWave.Wave3D.compute_vecv_abs_yz_left!

MaxwellWave.Wave3D.compute_vecv_abs_yz_right!

MaxwellWave.Wave3D.update_vecu!

MaxwellWave.Wave3D.update_vecu_abs!

MaxwellWave.Wave3D.update_vecu_abs_yz_left!

MaxwellWave.Wave3D.update_vecv_nabla2!

MaxwellWave.Wave3D.update_vecv_sigma!

MaxwellWave.Wave3D.update_vecv_varepsilon!

MaxwellWave.Wave3DnonlinearChi3.compute_alpha!

MaxwellWave.Wave3DnonlinearChi3.update_vecv_chi3!