Wave 3D

Auxiliary

Introduction to MaxwellWave

Maxwell Wave

Derivation Partial Differentaial Equation

Boundary Condition

Results

Conclusion and Possible Extensions

Running the Project Examples

Wave 2D

Wave 3D

Nonlinear Chi3 Wave 3D

MaxwellWave.Wave3D.update_vecu! — Function

@parallel_indices (ix, iy, iz) function update_vecu!(ux, uy, uz, vx, vy, vz, dt)

Computes the update rule for the vector field through the equation:

\[\vec{u}^{j+1} = \vec{u}^{j} + dt*\vec{v}^{j+\frac{1}{2}}\]

Important to call the kernel with size(u)-2

MaxwellWave.Wave3D.update_vecv_nabla2! — Function

@parallel_indices (ix, iy, iz) function update_vecv_nabla2!(ux, uy, uz, vx, vy, vz, dt, _dx2, _dy2, _dz2, alpha)

Computes the update rule for the velocity of the vector field through the equation:

\[\vec{v}^{j+\frac{1}{2}} = \vec{v}^{j-\frac{1}{2}} + dt*\alpha\left(\vec{r}\right) \nabla^2 \vec{u}^{j}\]

MaxwellWave.Wave3D.update_vecv_sigma! — Function

@parallel_indices (ix, iy, iz) function update_vecv_sigma!(ux, uy, uz, vx, vy, vz, dt, beta, gamma)

Computes the update rule for the velocity of the vector field through the equation:

\[\vec{v}^{j+\frac{1}{2}} = \vec{v}^{j-\frac{1}{2}} + dt*\beta\left(\vec{r}\right) \vec{u}^{j} + dt*\gamma\left(\vec{r}\right) \vec{v}^{j-\frac{1}{2}}\]

Important to call the kernel with size(u)-2

MaxwellWave.Wave3D.update_vecv_varepsilon! — Function

@parallel_indices (ix, iy, iz) function update_vecv_varepsilon!(ux, uy, uz, vx, vy, vz, dt, _dx_2, _dy_2, _dz_2, alpha, etax, etay, etaz)

Computes the update rule for the velocity of the vector field through the equation:

\[\vec{v}^{j+\frac{1}{2}} = \vec{v}^{j-\frac{1}{2}} + dt*\alpha\left(\vec{r}\right) \nabla \left(\vec{\eta}\left(\vec{r}\right) \cdot \vec{u}^{j} \right)\]

and discretized in the following way:

\[v_x^{j+\frac{1}{2}} \left[ix,iy,iz\right] = v_x^{j-\frac{1}{2}}\left[ix,iy,iz\right] + dt*\alpha\left[ix,iy,iz\right] \left( \left(\eta_x\left[ix+1,iy,iz\right] u_x^{j}\left[ix+1,iy,iz\right] - \eta_x\left[ix-1,iy,iz\right] u_x^{j}\left[ix-1,iy,iz\right]\right) + \left(\eta_y\left[ix+1,iy,iz\right] u_y^{j}\left[ix+1,iy,iz\right] - \eta_y\left[ix-1,iy,iz\right] u_y^{j}\left[ix-1,iy,iz\right]\right) + \left(\eta_z\left[ix+1,iy,iz\right] u_z^{j}\left[ix+1,iy,iz\right] - \eta_z\left[ix-1,iy,iz\right] u_z^{j}\left[ix-1,iy,iz\right]\right)\right)/dx/2\]

\[v_y^{j+\frac{1}{2}} \left[ix,iy,iz\right] = v_y^{j-\frac{1}{2}}\left[ix,iy,iz\right] + dt*\alpha\left[ix,iy,iz\right] \left( \left(\eta_x\left[ix,iy+1,iz\right] u_x^{j}\left[ix,iy+1,iz\right] - \eta_x\left[ix,iy-1,iz\right] u_x^{j}\left[ix,iy-1,iz\right]\right) + \left(\eta_y\left[ix,iy+1,iz\right] u_y^{j}\left[ix,iy+1,iz\right] - \eta_y\left[ix,iy-1,iz\right] u_y^{j}\left[ix,iy-1,iz\right]\right) + \left(\eta_z\left[ix,iy+1,iz\right] u_z^{j}\left[ix,iy+1,iz\right] - \eta_z\left[ix,iy-1,iz\right] u_z^{j}\left[ix,iy-1,iz\right]\right)\right)/dy/2\]

\[v_z^{j+\frac{1}{2}} \left[ix,iy,iz\right] = v_z^{j-\frac{1}{2}}\left[ix,iy,iz\right] + dt*\alpha\left[ix,iy,iz\right] \left( \left(\eta_x\left[ix,iy,iz+1\right] u_x^{j}\left[ix,iy,iz+1\right] - \eta_x\left[ix,iy,iz-1\right] u_x^{j}\left[ix,iy,iz-1\right]\right) + \left(\eta_y\left[ix,iy,iz+1\right] u_y^{j}\left[ix,iy,iz+1\right] - \eta_y\left[ix,iy,iz-1\right] u_y^{j}\left[ix,iy,iz-1\right]\right) + \left(\eta_z\left[ix,iy,iz+1\right] u_z^{j}\left[ix,iy,iz+1\right] - \eta_z\left[ix,iy,iz-1\right] u_z^{j}\left[ix,iy,iz-1\right]\right)\right)/dz/2\]

MaxwellWave.Wave3D.update_vecu_abs_yz_left! — Function

@parallel_indices (ix, iy, iz) function update_vecu_abs_yz_left!(ux, uy, uz, vx, vy, vz, dt)

Computes the update rule for the vector field through the equation:

\[\vec{u}^{j+1} = \vec{u}^{j} + dt*\vec{v}^{j+\frac{1}{2}}\]

But the yz-plain at x=0 is a absorbing plaine. Important to call the kernel with size(u)-2 in y,z, but with size(u)-1 in x This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.compute_vecv_abs_yz_left! — Function

@parallel_indices (iy, iz) function update_vecv_abs_yz_left!(ux, uy, uz, vx, vy, vz, c)

Computes the update rule for the vector field through the equation:

\[\frac{\partial}{\partial t}\vec{E}\left(\vec{r},t\right) + c \left(\vec{n} \cdot \nabla \right)\vec{E}\left(\vec{r},t\right)\]

The yz-plain at x=0 is a absorbing plaine. Important to call the kernel with size(u)-2 in y,z This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.compute_vecv_abs_yz_right! — Function

@parallel_indices (iy, iz) function compute_vecv_abs_yz_right!(ux, uy, uz, vx, vy, vz, c, _dx)

Computes the update rule for the vector field through the equation:

\[\frac{\partial}{\partial t}\vec{E}\left(\vec{r},t\right) + c \left(\vec{n} \cdot \nabla \right)\vec{E}\left(\vec{r},t\right)\]

The yz-plain at x=end is a absorbing plaine. Important to call the kernel with size(u)-2 in y,z This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.compute_vecv_abs_xz_left! — Function

@parallel_indices (ix, iz) function update_vecv_abs_xz_left!(ux, uy, uz, vx, vy, vz, c)

Computes the update rule for the vector field through the equation:

\[\frac{\partial}{\partial t}\vec{E}\left(\vec{r},t\right) + c \left(\vec{n} \cdot \nabla \right)\vec{E}\left(\vec{r},t\right)\]

The xz-plain at y=0 is a absorbing plaine. Important to call the kernel with size(u)-2 in x,z This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.compute_vecv_abs_xz_right! — Function

@parallel_indices (ix, iz) function compute_vecv_abs_xz_right!(ux, uy, uz, vx, vy, vz, c, _dx)

Computes the update rule for the vector field through the equation:

\[\frac{\partial}{\partial t}\vec{E}\left(\vec{r},t\right) + c \left(\vec{n} \cdot \nabla \right)\vec{E}\left(\vec{r},t\right)\]

The xz-plain at y=end is a absorbing plaine. Important to call the kernel with size(u)-2 in x,z This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.compute_vecv_abs_xy_left! — Function

@parallel_indices (ix, iy) function compute_vecv_abs_xy_left!(ux, uy, uz, vx, vy, vz, c, _dx)

Computes the update rule for the vector field through the equation:

\[\frac{\partial}{\partial t}\vec{E}\left(\vec{r},t\right) + c \left(\vec{n} \cdot \nabla \right)\vec{E}\left(\vec{r},t\right)\]

The xy-plain at yzl the kernel with size(u)-2 in x,y This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.compute_vecv_abs_xy_right! — Function

@parallel_indices (ix, iy) function compute_vecv_abs_xy_right!(ux, uy, uz, vx, vy, vz, c, _dx)

Computes the update rule for the vector field through the equation:

\[\frac{\partial}{\partial t}\vec{E}\left(\vec{r},t\right) + c \left(\vec{n} \cdot \nabla \right)\vec{E}\left(\vec{r},t\right)\]

The xy-plain at z=end is a absorbing plaine. Important to call the kernel with size(u)-2 in x,y This kernel is only allowed to be called from the ranks located at this domain boundary

MaxwellWave.Wave3D.update_vecu_abs! — Function

@parallel_indices (ix, iy, iz) function update_vecu_abs(ux, uy, uz, vx, vy, vz, dt, neighbors_x, neighbors_y, neighbors_z, abs_x, abs_y, abs_z)

Computes the update rule for the vector field through the equation:

\[\vec{u}^{j+1} = \vec{u}^{j} + dt*\vec{v}^{j+\frac{1}{2}}\]

More general kernel, where depending on arguments it is decided if boundary is absorbing. neighborsi gives the information if rank lies on left/right boundary and absi if this boundary is absorbing.

MaxwellWave.Auxiliary.load_array

MaxwellWave.Auxiliary.save_array

MaxwellWave.Wave2D.update_u!

MaxwellWave.Wave2D.update_v_nabla2!

MaxwellWave.Wave3D.compute_vecv_abs_xy_left!

MaxwellWave.Wave3D.compute_vecv_abs_xy_right!

MaxwellWave.Wave3D.compute_vecv_abs_xz_left!

MaxwellWave.Wave3D.compute_vecv_abs_xz_right!

MaxwellWave.Wave3D.compute_vecv_abs_yz_left!

MaxwellWave.Wave3D.compute_vecv_abs_yz_right!

MaxwellWave.Wave3D.update_vecu!

MaxwellWave.Wave3D.update_vecu_abs!

MaxwellWave.Wave3D.update_vecu_abs_yz_left!

MaxwellWave.Wave3D.update_vecv_nabla2!

MaxwellWave.Wave3D.update_vecv_sigma!

MaxwellWave.Wave3D.update_vecv_varepsilon!

MaxwellWave.Wave3DnonlinearChi3.compute_alpha!

MaxwellWave.Wave3DnonlinearChi3.update_vecv_chi3!