Nonlinear Chi3 Wave 3D

Auxiliary

Introduction to MaxwellWave

Maxwell Wave

Derivation Partial Differentaial Equation

Boundary Condition

Results

Conclusion and Possible Extensions

Running the Project Examples

Wave 2D

Wave 3D

Nonlinear Chi3 Wave 3D

MaxwellWave.Wave3DnonlinearChi3.update_vecv_chi3! — Function

@parallel_indices (ix, iy, iz) function update_vecv_chi3!(ux, uy, uz, vx, vy, vz, dt, _dx, _dy, _dz, _dx_2, _dy_2, _dz_2, alpha, delta, omega)

Computes the update rule for the velocity of the vector field through the equation:

\[\vec{v}^{j+\frac{1}{2}} = \vec{v}^{j-\frac{1}{2}} + dt*\alpha\left(\vec{r}\right) \nabla \left(\delta \left(\vec{r}\right) \nabla \cdot \left( \omega \left(\vec{r}\right) \left\lvert \vec{u}^{j} \right\rvert^2 \vec{u}^{j} \right) \right)\]

and discretized in the following way:

\[v_x^{j+\frac{1}{2}}\left[ix,iy,iz\right] = v_x^{j-\frac{1}{2}}\left[ix,iy,iz\right] + dt*\alpha\left[ix,iy,iz\right] \left( \delta \left[ix+1,iy,iz\right] \left( \left(\omega \left[ix+1,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix+1,iy,iz\right] \right\rvert^2 u_x^j \left[ix+1,iy,iz\right] - \omega \left[ix,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz\right] \right\rvert^2 u_x^j\left[ix,iy,iz\right]\right)/dx + \left(\omega \left[ix+1,iy+1,iz\right] \left\lvert \vec{u}^{j}\left[ix+1,iy+1,iz\right] \right\rvert^2 u_y^j\left[ix+1,iy+1,iz\right] - \omega \left[ix+1,iy-1,iz\right] \left\lvert \vec{u}^{j}\left[ix+1,iy-1,iz\right] \right\rvert^2 u_y^j\left[ix+1,iy-1,iz\right]\right)/dy/2+ \left(\omega \left[ix+1,iy,iz+1\right] \left\lvert \vec{u}^{j}\left[ix+1,iy,iz+1\right] \right\rvert^2 u_z^j\left[ix+1,iy,iz+1\right]- \omega \left[ix+1,iy,iz-1\right] \left\lvert \vec{u}^{j}\left[ix+1,iy,iz-1\right] \right\rvert^2 u_z^j\left[ix+1,iy,iz-1\right]\right)/dz/2\right) + \delta \left[ix-1,iy,iz\right] \left( \left(\omega \left[ix,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz\right] \right\rvert^2 u_x^j\left[ix,iy,iz\right] - \omega \left[ix-1,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix-1,iy,iz\right] \right\rvert^2 u_x^j\left[ix-1,iy,iz\right]\right)/dx + \left(\omega \left[ix-1,iy+1,iz\right] \left\lvert \vec{u}^{j}\left[ix-1,iy+1,iz\right] \right\rvert^2 u_y^j\left[ix-1,iy+1,iz\right] - \omega \left[ix-1,iy-1,iz\right] \left\lvert \vec{u}^{j}\left[ix-1,iy-1,iz\right] \right\rvert^2 u_y^j\left[ix-1,iy-1,iz\right]\right)/dy/2 + \left(\omega \left[ix-1,iy,iz+1\right] \left\lvert \vec{u}^{j}\left[ix-1,iy,iz+1\right] \right\rvert^2 u_z^j\left[ix-1,iy,iz+1\right]- \omega \left[ix-1,iy,iz-1\right] \left\lvert \vec{u}^{j}\left[ix-1,iy,iz-1\right] \right\rvert^2 u_z^j\left[ix-1,iy,iz-1\right]\right)/dz/2\right) \right)/dx/2\]

\[v_y^{j+\frac{1}{2}}\left[ix,iy,iz\right] = v_y^{j-\frac{1}{2}}\left[ix,iy,iz\right] + dt*\alpha\left[ix,iy,iz\right] \left( \delta \left[ix,iy+1,iz\right] \left( \left(\omega \left[ix+1,iy+1,iz\right] \left\lvert \vec{u}^{j}\left[ix+1,iy+1,iz\right] \right\rvert^2 u_x^j \left[ix+1,iy+1,iz\right] - \omega \left[ix-1,iy+1,iz\right] \left\lvert \vec{u}^{j}\left[ix-1,iy+1,iz\right] \right\rvert^2 u_x^j\left[ix-1,iy+1,iz\right]\right)/dx/2 + \left(\omega \left[ix,iy+1,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy+1,iz\right] \right\rvert^2 u_y^j\left[ix,iy+1,iz\right] - \omega \left[ix,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz\right] \right\rvert^2 u_y^j\left[ix,iy,iz\right]\right)/dy+ \left(\omega \left[ix,iy+1,iz+1\right] \left\lvert \vec{u}^{j}\left[ix,iy+1,iz+1\right] \right\rvert^2 u_z^j\left[ix,iy+1,iz+1\right]- \omega \left[ix,iy+1,iz-1\right] \left\lvert \vec{u}^{j}\left[ix,iy+1,iz-1\right] \right\rvert^2 u_z^j\left[ix,iy+1,iz-1\right]\right)/dz/2\right) + \delta \left[ix,iy-1,iz\right] \left( \left(\omega \left[ix+1,iy-1,iz\right] \left\lvert \vec{u}^{j}\left[ix+1,iy-1,iz\right] \right\rvert^2 u_x^j\left[ix+1,iy-1,iz\right] - \omega \left[ix-1,iy-1,iz\right] \left\lvert \vec{u}^{j}\left[ix-1,iy-1,iz\right] \right\rvert^2 u_x^j\left[ix-1,iy-1,iz\right]\right)/dx/2 + \left(\omega \left[ix,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz\right] \right\rvert^2 u_y^j\left[ix,iy,iz\right] - \omega \left[ix,iy-1,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy-1,iz\right] \right\rvert^2 u_y^j\left[ix,iy-1,iz\right]\right)/dy + \left(\omega \left[ix,iy-1,iz+1\right] \left\lvert \vec{u}^{j}\left[ix,iy-1,iz+1\right] \right\rvert^2 u_z^j\left[ix,iy-1,iz+1\right]- \omega \left[ix,iy-1,iz-1\right] \left\lvert \vec{u}^{j}\left[ix,iy-1,iz-1\right] \right\rvert^2 u_z^j\left[ix,iy-1,iz-1\right]\right)/dz/2\right) \right)/dy/2\]

\[v_z^{j+\frac{1}{2}}\left[ix,iy,iz\right] = v_z^{j-\frac{1}{2}}\left[ix,iy,iz\right] + dt*\alpha\left[ix,iy,iz\right] \left( \delta \left[ix,iy,iz+1\right] \left( \left(\omega \left[ix+1,iy,iz+1\right] \left\lvert \vec{u}^{j}\left[ix+1,iy,iz+1\right] \right\rvert^2 u_x^j \left[ix+1,iy,iz+1\right] - \omega \left[ix-1,iy,iz+1\right] \left\lvert \vec{u}^{j}\left[ix-1,iy,iz+1\right] \right\rvert^2 u_x^j\left[ix-1,iy,iz+1\right]\right)/dx/2 + \left(\omega \left[ix,iy+1,iz+1\right] \left\lvert \vec{u}^{j}\left[ix,iy+1,iz+1\right] \right\rvert^2 u_y^j\left[ix,iy+1,iz+1\right] - \omega \left[ix,iy-1,iz+1\right] \left\lvert \vec{u}^{j}\left[ix,iy-1,iz+1\right] \right\rvert^2 u_y^j\left[ix,iy-1,iz+1\right]\right)/dy/2+ \left(\omega \left[ix,iy,iz+1\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz+1\right] \right\rvert^2 u_z^j\left[ix,iy,iz+1\right]- \omega \left[ix,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz\right] \right\rvert^2 u_z^j\left[ix,iy,iz\right]\right)/dz\right) + \delta \left[ix,iy,iz-1\right] \left( \left(\omega \left[ix+1,iy,iz-1\right] \left\lvert \vec{u}^{j}\left[ix+1,iy,iz-1\right] \right\rvert^2 u_x^j\left[ix+1,iy,iz-1\right] - \omega \left[ix-1,iy,iz-1\right] \left\lvert \vec{u}^{j}\left[ix-1,iy,iz-1\right] \right\rvert^2 u_x^j\left[ix-1,iy,iz-1\right]\right)/dx/2 + \left(\omega \left[ix,iy+1,iz-1\right] \left\lvert \vec{u}^{j}\left[ix,iy+1,iz-1\right] \right\rvert^2 u_y^j\left[ix,iy+1,iz-1\right] - \omega \left[ix,iy-1,iz-1\right] \left\lvert \vec{u}^{j}\left[ix,iy-1,iz-1\right] \right\rvert^2 u_y^j\left[ix,iy-1,iz-1\right]\right)/dy/2 + \left(\omega \left[ix,iy,iz\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz\right] \right\rvert^2 u_z^j\left[ix,iy,iz\right]- \omega \left[ix,iy,iz-1\right] \left\lvert \vec{u}^{j}\left[ix,iy,iz-1\right] \right\rvert^2 u_z^j\left[ix,iy,iz-1\right]\right)/dz\right) \right)/dz/2\]

Where we have applied a central difference formula for the outer derivative. For the inner derivates, we use the central or forward/backward formula depending on if the derivative is in the same direction as the outer derivative. We use the forward/backward formula to arrive at a local nearest neighbor stencil.

MaxwellWave.Wave3DnonlinearChi3.compute_alpha! — Function

@parallel_indices (ix, iy, iz) function compute_alpha!(alpha, ux, uy, uz, _mu0, _epsilon0, chi3, epsilon)

Computes the new coefficient alpha in every step:

\[alpha = \frac{1}{\mu_0 \varepsilon_0 \varepsilon_r\left(\vec{r}\right) + \mu_0\varepsilon_0 \chi^3\left(\vec{r}\right) \left\lvert \vec{E}\left(\vec{r},t\right)\right\rvert^2}\]

MaxwellWave.Auxiliary.load_array

MaxwellWave.Auxiliary.save_array

MaxwellWave.Wave2D.update_u!

MaxwellWave.Wave2D.update_v_nabla2!

MaxwellWave.Wave3D.compute_vecv_abs_xy_left!

MaxwellWave.Wave3D.compute_vecv_abs_xy_right!

MaxwellWave.Wave3D.compute_vecv_abs_xz_left!

MaxwellWave.Wave3D.compute_vecv_abs_xz_right!

MaxwellWave.Wave3D.compute_vecv_abs_yz_left!

MaxwellWave.Wave3D.compute_vecv_abs_yz_right!

MaxwellWave.Wave3D.update_vecu!

MaxwellWave.Wave3D.update_vecu_abs!

MaxwellWave.Wave3D.update_vecu_abs_yz_left!

MaxwellWave.Wave3D.update_vecv_nabla2!

MaxwellWave.Wave3D.update_vecv_sigma!

MaxwellWave.Wave3D.update_vecv_varepsilon!

MaxwellWave.Wave3DnonlinearChi3.compute_alpha!

MaxwellWave.Wave3DnonlinearChi3.update_vecv_chi3!